Move PortUtil helper methods into TestHelper.
